authorization implemented

+++ notes-app-back-end/src/api/notes/handler.js
@@ -16,9 +16,9 @@ class NotesHandler {
         try {
             this._validator.validateNotePayload(request.payload);
             const {title = 'untitled', body, tags} = request.payload;
+            const { id: credentialId } = request.auth.credentials;
 
-            this._service.addNote({title, body, tags});
-            const noteId = await this._service.addNote({title, body, tags});
+            const noteId = await this._service.addNote({title, body, tags, owner: credentialId});
     
             const response = h.response({
                 status: 'success',
@@ -50,8 +50,9 @@ class NotesHandler {
         }
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request, h) {
+        const { id: credentialId } = request.auth.credentials;
+        const notes = await this._service.getNotes(credentialId);
         return {
             status: 'success',
             data: {
@@ -62,7 +63,10 @@ class NotesHandler {
 
     async getNoteByIdHandler(request, h) {
         try {
-            const {id} = request.params;
+            const { id } = request.params;
+            const { id: credentialId } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
             const note = await this._service.getNoteById(id);
             return {
                 status: 'success',
@@ -94,8 +98,10 @@ class NotesHandler {
     async putNoteByIdHandler(request, h) {
         try {
             this._validator.validateNotePayload(request.payload);
-            const {id} = request.params;
+            const { id } = request.params;
+            const { id: credentialId } = request.auth.credentials;
 
+            await this._service.verifyNoteOwner(id, credentialId);
             await this._service.editNoteById(id, request.payload);
 
             return {
@@ -125,7 +131,10 @@ class NotesHandler {
 
     async deleteNoteByIdHandler(request, h) {
         try {
-            const {id} = request.params;
+            const { id } = request.params;
+            const { id: credentialId } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
             await this._service.deleteNoteById(id);
 
             return {
